fix(icon): rtl
