--- src/packages/icon/demo.scss
@@ -1,30 +1,0 @@
-.nut-cell > .nutui-iconfont {
-  margin-right: 10px;
-}
-
-.h5-ul {
-  display: flex;
-  flex-wrap: wrap;
-  padding: 0;
-  width: 100%;
-}
-
-.h5-ul .h5-li {
-  flex: 0 0 25%;
-  max-width: 25%;
-  height: 40px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.h5-ul .h5-li .h5-span {
-  height: 40px;
-  font-size: 12px;
-  text-align: center;
-}
-
-.h5-ul .h5-li .h5-span .nutui-iconfont {
-  margin: 16px 0 16px;
-}
